:pencil2: rename scripts and classes to comply with naming convention - :bug: ignore missing selector in AdvancedSelectionManager, just a problem during loading of scenes not in game
