+++ Assets/Scripts/Colision/Voxeldoshake.cs
@@ -9,7 +9,7 @@ public class Voxeldoshake : bycolision
 
 {
     public override void doaction(GameObject gobj, Collider otherCollider) {
-        otherCollider.transform.GetComponent<vertex_waves_shader_logic>().bycollision();
+        otherCollider.transform.GetComponent<VertexWavesShaderLogic>().DoShake();
         if (sound is not null)
         {
             AudioSource.PlayClipAtPoint(sound, otherCollider.transform.position);
